requirements.txt
numpy==1.23.2
pygame==2.1.2
Pillow==9.2.0
Cython==0.29.32
PyGLM==2.5.7
pytest[tests]==7.1.2